version 2 - length works...mostly
Tested with a length of 7. Tested on every combination that didn't use the checkbox uppercase. When I did this every password generated came back with a length of 8. When uppercase and any other check box was selected the length was however many checkboxes were selected. I solved this probem and it turned out that I had 3 s's instead of 2 in a word.

Expected: Length should work when any combination of checkboxes is ticked
Boundary: Length on maximum and minimum should work
Invald: If no length or a length higher or lower than the maximum(20) and minium (6) length is entered submit an error message ( that bit works, tried and tested on 5,100, 21 and nothing)