Version 3 - fixed errors
Expected: Length and output works for all combinations
Boundary: Length works for the maximum (20) and minimum(6) length - tried and tested on both 6 and 20
Invalid: If no length is selected then an error message is outputted. If a length above the maximum is wanted then an error message is outputted and if a length less than 6 another error message is outputted. tried an tested with nothing, 3, 5, 21, and 500.